397 - Remove useExperiments hook and use getServerSideProps for experiment data fetching (#398)
* (delete) remove useExperiments hook and add new helpers/properties in config/readable

* (edit) replace runtime data fetching with getServerSideProps in [accession_code] and [title] (previously [experiment]) and adjust codebase accordingly

* (minor) clean up the experiment check in useEffect in AccessionCode

* (edit) remove experiment from the depedency array (not necessary) and replace early return with JSX ternary for clarity in Experiment, improve comment and fix a typo

* (edit) use getServerSideProps for redirect in [accesstion_code], and move formatting for platform names and technology to SearchCardAction and SearchCardMeta components

* (edit) extract the experiment detail UI and the experiment samples table from the [experiment] page component for better readability

* (edit) revise the InformationItem css styles and add the fallback and forceFallback props

* (edit) pass the entire experiment to helpers and handle the edge cases within them, and fix typo in [accession_code]

* (edit) replace id with accession_code
